fixed pom files so we have the right dependecies

+++ src/main/java/facade/FacadeCurrent.java
@@ -21,7 +21,7 @@ public class FacadeCurrent
 
     EntityManagerFactory emf;
 
-    public Person getPerson(int id) //Kunne alternativt være telefonnummer
+    public Person getPerson(long id) //Kunne alternativt være telefonnummer
             
     {
         EntityManager em = emf.createEntityManager();
